scripts/visualize/requirements.txt
numpy
open3d
Pillow
scipy
tqdm
